Refactor : css 변수 사용 방법 수정
