+++ src/jetty-chart/src/bars/normal-bar/nomal-bar.js
@@ -164,7 +164,7 @@ const NormalBar = ({
       legendSettings={result.legendSettings}
       animationSettings={result.animationSettings}
     >
-      <g transform={calculateWarpperTransform({ horizontal, reverse, innerMargin, padding })}>
+      <g transform={calculateWarpperTransform({ horizontal, reverse, innerMargin, padding })} className={styles.container}>
         {data.map((d, idx) => {
           const nowData = { ...d };
