+++ src/jetty-chart/src/bars/normal-bar/normal-bar.module.css
@@ -1,4 +1,4 @@
-svg {
+.container {
   --bar-from: 0, 0;
   --bar-to: 0, 0;
   --text-from: 0, 0;
