+++ src/jetty-chart/src/components/x-axis/x-axis-label.module.css
@@ -1,4 +1,4 @@
-svg {
+.container {
   --animation-duration: 0;
   --animation-timing-function: "linear";
   --animation-delay: "0s";
