+++ src/jetty-chart/src/components/y-axis/draw-y-axis-grid-line.js
@@ -38,7 +38,7 @@ export const DrawYAxisGridLine = ({
   }
 
   return (
-    <g>
+    <g className={styles.container}>
       {yAxis.map((c, idx) => {
         if (!showTopScope && (idx === 0 || idx === yAxis.length - 1) && c !== 0) {
           return;
