+++ src/jetty-chart/src/components/y-axis/y-axis-grid-line.module.css
@@ -1,4 +1,4 @@
-svg {
+.container {
   --line-width: "0px";
   --line-offset: "0px";
   --animation-duration: 0;
