+++ src/jetty-chart/src/components/y-axis/y-axis-label.module.css
@@ -1,4 +1,4 @@
-svg {
+.container {
   --animation-duration: 0;
   --animation-timing-function: "linear";
   --animation-delay: "0s";
