+++ src/jetty-chart/src/lines/bump/bump.js
@@ -252,7 +252,7 @@ const BumpChart = ({
       legendSettings={result.legendSettings}
       animationSettings={result.animationSettings}
     >
-      <g transform={horizontal ? `translate(0,${padding})` : `translate(${padding})`}>
+      <g transform={horizontal ? `translate(0,${padding})` : `translate(${padding})`} className={styles.container}>
         {enableArea &&
           areaPathArray.map((d, idx) => {
             return (
