+++ src/jetty-chart/src/lines/bump/bump.module.css
@@ -1,4 +1,4 @@
-svg {
+.container {
   --line-width: "0px";
   --line-height: "0px";
   --line-length: "0px";
