+++ src/jetty-chart/src/maps/normal-map/normal-map.module.css
@@ -1,12 +1,12 @@
-path:hover{
-    transform: scale(1.0005);
-    transform-origin: 50% 50%;
-    stroke: white;
-    stroke-width: 2;
-    fill-opacity: 1;
-}
-g:hover{
-    stroke-width: 2;
-    fill-opacity: 1; 
+.path:hover {
+  transform: scale(1.0005);
+  transform-origin: 50% 50%;
+  stroke: white;
+  stroke-width: 2;
+  fill-opacity: 1;
 }
 
+.group:hover {
+  stroke-width: 2;
+  fill-opacity: 1;
+}
